moving test input dict to location that should work better with python2
